+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Prevent potential intercepting output of other ant tasks (#3)
+### 1.3.0 (2017-03-20)
+* Prevent other tasks output interception (#3)
 * Change console reporting format (to match gradle-quality-plugin format):
     - Each error split to source and code lines followed by empty line
     - IDE will be able to show link for class reference (clickable) when line number available